+++ src/main/java/net/witcher_rpg/client/WitcherClient.java
@@ -9,6 +9,7 @@ import net.minecraft.client.particle.*;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
 import net.spell_engine.api.render.CustomModels;
 import net.witcher_rpg.client.effect.AxiiParticles;
+import net.witcher_rpg.client.effect.QuenParticles;
 import net.witcher_rpg.client.entity.YrdenRenderer;
 import net.witcher_rpg.client.particle.Particles;
 import net.witcher_rpg.effect.Effects;
@@ -34,6 +35,7 @@ public class WitcherClient implements ClientModInitializer {
         ParticleFactoryRegistry.getInstance().register(Particles.YRDEN_CLOUD, DragonBreathParticle.Factory::new);
 
         CustomParticleStatusEffect.register(Effects.AXII, new AxiiParticles(1));
+        CustomParticleStatusEffect.register(Effects.QUEN_SHIELD, new QuenParticles(1));
 
         EntityRendererRegistry.register(YrdenEntity.ENTITY_TYPE,YrdenRenderer::new);
     }