+++ src/main/java/net/witcher_rpg/client/effect/QuenParticles.java
@@ -1,0 +1,24 @@
+package net.witcher_rpg.client.effect;
+
+import net.minecraft.entity.LivingEntity;
+import net.spell_engine.api.effect.CustomParticleStatusEffect;
+import net.spell_engine.api.spell.ParticleBatch;
+import net.spell_engine.particle.ParticleHelper;
+
+public class QuenParticles implements CustomParticleStatusEffect.Spawner {
+    private final ParticleBatch particles;
+
+    public QuenParticles(int particleCount) {
+        this.particles = new ParticleBatch(
+                "spell_engine:electric_arc_a",
+                ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.CENTER,
+                null, particleCount, 0.01F, 0.1F, 0);
+    }
+
+    @Override
+    public void spawnParticles(LivingEntity livingEntity, int amplifier) {
+        var scaledParticles = new ParticleBatch(particles);
+        scaledParticles.count *= (1);
+        ParticleHelper.play(livingEntity.getWorld(), livingEntity, scaledParticles);
+    }
+}