+++ src/main/java/net/witcher_rpg/config/EffectsConfig.java
@@ -6,8 +6,9 @@ public class EffectsConfig {
     public float aerondight_spell_crit_damage_increase_per_stack = 0.05f;
     public float magical_trap_speed_reduction = -0.10f;
     public float axii_damage_increase = 0.250F;
-    public float adrenaline_attack_increase_per_stack = 0.10F;
-    public float adrenaline_sign_intensity_increase_per_stack = 0.05F;
     public float quen_reflect_damage = 0.50F;
-    public float adrenaline_attribute_increase_per_stack = 0.001F;
+    public float adrenaline_attack_increase_per_stack = 0.025F;
+    public float adrenaline_sign_intensity_increase_per_stack = 0.025F;
+    public int adrenaline_max_amplifier = 10;
+    public float adrenaline_attribute_increase_per_stack = 0.01F;
 }