+++ src/main/java/net/witcher_rpg/item/weapon/AerondightRelictItem.java
@@ -6,6 +6,7 @@ import net.minecraft.block.Blocks;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
@@ -43,13 +44,14 @@ public class AerondightRelictItem extends SpellSwordItem {
         int charge_incamp =1;
 
         if(target.isUndead()){
-            target.damage(target.getDamageSources().magic(), 2.0f);
+            float attack = (float) attacker.getAttributeValue(EntityAttributes.GENERIC_ATTACK_DAMAGE);
+            target.damage(target.getDamageSources().magic(), attack * 0.30F);
             player.addEnchantedHitParticles(target);
         }
         increaseHiddenEffectLevel(player, Effects.AERONDIGHT_CHARGE,charge_duration,charge_incamp,charge_amplifier_max);
         int currentAmplifier = player.getStatusEffect(Effects.AERONDIGHT_CHARGE).getAmplifier();
         if(currentAmplifier == charge_amplifier_max){
-                target.damage(target.getDamageSources().magic(), 5.0f);
+                target.damage(target.getDamageSources().magic(), 1.0f);
                 player.addEnchantedHitParticles(target);
             }
         stack.damage(1, attacker, (e)->{