+++ src/main/java/net/witcher_rpg/item/weapon/WintersBladeRelicItem.java
@@ -7,6 +7,10 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.item.ToolMaterial;
 import net.more_rpg_classes.effect.MRPGCEffects;
 import net.spell_engine.api.item.weapon.SpellSwordItem;
+import net.witcher_rpg.effect.Effects;
+
+import static net.more_rpg_classes.util.CustomMethods.increaseAmpByChance;
+import static net.witcher_rpg.WitcherClassMod.effectsConfig;
 
 public class WintersBladeRelicItem extends SpellSwordItem {
     public WintersBladeRelicItem(ToolMaterial material, Settings settings) {
@@ -16,14 +20,11 @@ public class WintersBladeRelicItem extends SpellSwordItem {
     //10% chance to frost
     @Override
     public boolean postHit(ItemStack stack, LivingEntity target, LivingEntity attacker) {
-        int frosted_duration = 100;
+        int frosted_duration = 200;
         int frosted_chance = 10;
-        int randomrange_freeze = (int) ((Math.random() * (1 + frosted_chance)) + 1);
+        increaseAmpByChance(target, MRPGCEffects.FROSTED,frosted_duration,0,9,frosted_chance);
 
-        if (randomrange_freeze >= frosted_duration ) {
-            target.addStatusEffect(new StatusEffectInstance(MRPGCEffects.FROSTED, frosted_duration));
 
-        }
         stack.damage(1, attacker, (e)->{
             e.sendEquipmentBreakStatus(EquipmentSlot.MAINHAND);
         });